small_vector::size() proper cast.

+++ include/chobo/small_vector.hpp
@@ -509,7 +509,7 @@ public:
 
     size_t size() const noexcept
     {
-        return m_end - m_begin;
+        return static_cast<size_t>(m_end - m_begin);
     }
 
     size_t max_size() const noexcept
